+++ PostProcessing/Runtime/PostProcessResources.cs
@@ -15,6 +15,8 @@ namespace UnityEngine.Rendering.PostProcessing
             public Shader bloom;
             public Shader copy;
             public Shader copyStd;
+            public Shader copyStdFromTexArray;
+            public Shader copyStdFromDoubleWide;
             public Shader discardAlpha;
             public Shader depthOfField;
             public Shader finalPass;