+++ PostProcessing/Runtime/PostProcessVolume.cs
@@ -55,7 +55,11 @@ namespace UnityEngine.Rendering.PostProcessing
     /// A post-process volume component holding a post-process profile.
     /// </summary>
     /// <seealso cref="RuntimeUtilities.DestroyVolume"/>
+#if UNITY_2018_3_OR_NEWER
+    [ExecuteAlways]
+#else
     [ExecuteInEditMode]
+#endif
     [AddComponentMenu("Rendering/Post-process Volume", 1001)]
     public sealed class PostProcessVolume : MonoBehaviour
     {
@@ -219,7 +223,7 @@ namespace UnityEngine.Rendering.PostProcessing
             }
 #endif
 
-            var scale = transform.localScale;
+            var scale = transform.lossyScale;
             var invScale = new Vector3(1f / scale.x, 1f / scale.y, 1f / scale.z);
             Gizmos.matrix = Matrix4x4.TRS(transform.position, transform.rotation, scale);