+++ CHANGELOG.md
@@ -4,12 +4,61 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.18-preview]
+
+### Fixed
+- Camera viewport wasn't working properly when outputting directly to the backbuffer.
+- More improvements to VR support.
+- Compatibility fixes for 2017.1 to 2017.4.
+- Post-processing wouldn't work when loaded from an asset bundle.
+
+### Changed
+- Scriptable Render Pipelines should now call `PostProcessLayer.UpdateVolumeSystem(Camera, CommandBuffer)` at the beginning of the frame.
+
+## [2.0.17-preview]
+
+### Fixed
+- First pass at improving VR support.
+- Assert on Invalid LDR Lookup Texture size; added a check in the inspector for the user.
+- Improved performance on Unity 2019.1+ by avoiding unnecessary blits if no other image effect is active.
+- Use new ASTC enums on unity 2019.1+.
+
+## [2.0.16-preview]
+
+### Fixed
+- Grain shader compilation errors on some mobile GPUs.
+- Compilation issue with Unity 2019.1+ due to an internal API change.
+
+## [2.0.15-preview]
+
+### Fixed
+- Warning on `[ShaderIncludePath]` in 2018.3+.
+
+## [2.0.14-preview]
+
+### Fixed
+- Bloom flicker in single-pass double-wide stereo rendering.
+- Right eye bloom offset in single-pass double-wide stereo rendering.
+- If any parent of PostProcessingVolume has non-identity scale the Gizmo is rendered incorrectly.
+- Cleanup error when going back'n'forth between Builtins & Scriptable pipelines.
+
+### Changed
+- Use `ExecuteAlways` in 2018.3+ for better compatibility with "Prefab Mode".
+
+## [2.0.13-preview]
+
+### Fixed
+- Compilation issue with Unity 2019.1.
+- Screen-space reflection memory leak.
+
 ## [2.0.12-preview]
 
 ### Fixed
 - Ambient Occlusion could distort the screen on Android/Vulkan.
 - Warning about SettingsProvider in 2018.3.
 - Fixed issue with physical camera mode not working with post-processing.
+- Fixed thread group warning message on Metal and Intel Iris.
+- Fixed compatibility with versions pre-2018.2.
 
 ## [2.0.10-preview]